+++ app/src/main/java/com/vishalk/rssbstream/presentation/components/ChangelogBottomSheet.kt
@@ -136,7 +136,7 @@ fun ChangelogBottomSheet(
     modifier: Modifier = Modifier
 ) {
     val context = LocalContext.current
-    val changelogUrl = "https://github.com/theovilardo/PixelPlay/blob/master/CHANGELOG.md"
+    val changelogUrl = "https://github.com/vishalkdotcom/rssb-stream/blob/master/CHANGELOG.md"
 
     val fabCornerRadius = 16.dp