+++ app/src/main/java/com/vishalk/rssbstream/presentation/screens/AboutScreen.kt
@@ -302,7 +302,7 @@ fun AboutScreen(
                         horizontalAlignment = Alignment.CenterHorizontally
                     ) {
                         Text(
-                            text = "Pixel Play",
+                            text = "RSSB Stream",
                             style = MaterialTheme.typography.headlineSmall,
                             fontWeight = FontWeight.Bold
                         )
@@ -339,7 +339,7 @@ fun AboutScreen(
                             .padding(horizontal = 24.dp, vertical = 20.dp)
                     ) {
                         Text(
-                            text = "Thanks for using Pixel Play!",
+                            text = "Thanks for using RSSB Stream!",
                             style = MaterialTheme.typography.titleLarge,
                             fontWeight = FontWeight.Bold,
                             modifier = Modifier.padding(bottom = 8.dp)